// src/constants/enum.ts
export enum DAY_OF_WEEK {
  MONDAY = 1,
  TUESDAY = 2,
  WEDNESDAY = 3,
  THURSDAY = 4,
  FRIDAY = 5,
  SATURDAY = 6,
  SUNDAY = 7,
}

export enum SUBJECT_TYPE {
  LECTURE = 'LECTURE',
  LAB = 'LAB',
}

export enum DEGREE {
  Bachelor = 'BACHELOR',
  BachelorCon = 'BACHELOR_CONTINUE',
  BachelorInter = 'BACHELOR_INTER',
  Pundit = 'PUNDIT',
  PunditInter = 'PUNDIT_INTER',
}